Messages now slist_tpl => much faster when many messages are present
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@2816 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/message_stats_t.h
@@ -23,7 +23,7 @@ class message_stats_t : public gui_komponente_t
 private:
 	message_t *msg;
 	karte_t *welt;
-	unsigned last_count;
+	uint16 last_count;
 	sint32 message_selected;
 
 public:
